(fileQueue) Extract preprocessing orchestration into usePreprocessing hook
Move the getPreprocessingManager singleton, its manager.on(...) event
subscriptions (status-change / completed / error / file-not-found /
paused / resumed / cache-cleared / already-processed), the pause/resume
toasts, missing-file batching, hash checker, cache-priority push and the
all-done completion toast out of FileQueueModule and into a dedicated
fileQueue/usePreprocessing hook. Move-only: behavior unchanged. The shell
consumes the returned preprocessing state, status ref, pause flag and
countPreprocessed helper.
